requirements.txt
streamlit
langchain
langchain-groq
langchain-core
python-dotenv
requests
beautifulsoup4